+++ packages/perseus/src/widgets/interactive-graph.test.tsx
@@ -1,5 +1,7 @@
 import invariant from "tiny-invariant";
 
+import {clone} from "../../../../testing/object-utils";
+
 import InteractiveGraph, {shouldUseMafs} from "./interactive-graph";
 
 import type {
@@ -28,12 +30,9 @@ describe("InteractiveGraph.validate on a segment question", () => {
             ],
         });
 
-        const result = InteractiveGraph.widget.validate(guess, rubric, null);
+        const result = InteractiveGraph.widget.validate(guess, rubric);
 
-        expect(result).toEqual({
-            type: "invalid",
-            message: null,
-        });
+        expect(result).toHaveInvalidInput();
     });
 
     it("does not award points if guess.coords is wrong", () => {
@@ -56,14 +55,9 @@ describe("InteractiveGraph.validate on a segment question", () => {
             ],
         });
 
-        const result = InteractiveGraph.widget.validate(guess, rubric, null);
+        const result = InteractiveGraph.widget.validate(guess, rubric);
 
-        expect(result).toEqual({
-            type: "points",
-            earned: 0,
-            total: 1,
-            message: null,
-        });
+        expect(result).toHaveBeenAnsweredIncorrectly();
     });
 
     it("awards points if guess.coords is right", () => {
@@ -86,14 +80,9 @@ describe("InteractiveGraph.validate on a segment question", () => {
             ],
         });
 
-        const result = InteractiveGraph.widget.validate(guess, rubric, null);
+        const result = InteractiveGraph.widget.validate(guess, rubric);
 
-        expect(result).toEqual({
-            type: "points",
-            earned: 1,
-            total: 1,
-            message: null,
-        });
+        expect(result).toHaveBeenAnsweredCorrectly();
     });
 
     it("allows points of a segment to be specified in reverse order", () => {
@@ -116,14 +105,9 @@ describe("InteractiveGraph.validate on a segment question", () => {
             ],
         });
 
-        const result = InteractiveGraph.widget.validate(guess, rubric, null);
+        const result = InteractiveGraph.widget.validate(guess, rubric);
 
-        expect(result).toEqual({
-            type: "points",
-            earned: 1,
-            total: 1,
-            message: null,
-        });
+        expect(result).toHaveBeenAnsweredCorrectly();
     });
 
     it("does not modify the `guess` data", () => {
@@ -146,7 +130,7 @@ describe("InteractiveGraph.validate on a segment question", () => {
             ],
         });
 
-        InteractiveGraph.widget.validate(guess, rubric, null);
+        InteractiveGraph.widget.validate(guess, rubric);
 
         expect(guess.coords).toEqual([
             [
@@ -176,7 +160,7 @@ describe("InteractiveGraph.validate on a segment question", () => {
             ],
         });
 
-        InteractiveGraph.widget.validate(guess, rubric, null);
+        InteractiveGraph.widget.validate(guess, rubric);
 
         // Narrow the type of `rubric.correct` to segment graph; otherwise TS
         // thinks it might not have a `coords` property.
@@ -187,6 +171,138 @@ describe("InteractiveGraph.validate on a segment question", () => {
                 [0, 0],
             ],
         ]);
+    });
+});
+
+describe("InteractiveGraph.validate on a point question", () => {
+    it("marks the answer invalid if guess.coords is missing", () => {
+        const guess: PerseusGraphType = {type: "point"};
+        const rubric: PerseusInteractiveGraphRubric = createRubric({
+            type: "point",
+            coords: [[0, 0]],
+        });
+
+        const result = InteractiveGraph.widget.validate(guess, rubric);
+
+        expect(result).toHaveInvalidInput();
+    });
+
+    it("throws an exception if correct.coords is missing", () => {
+        // Characterization test: this might not be desirable behavior, but
+        // it's the current behavior as of 2024-09-25.
+        const guess: PerseusGraphType = {
+            type: "point",
+            coords: [[0, 0]],
+        };
+        const rubric: PerseusInteractiveGraphRubric = createRubric({
+            type: "point",
+        });
+
+        expect(() =>
+            InteractiveGraph.widget.validate(guess, rubric),
+        ).toThrowError();
+    });
+
+    it("does not award points if guess.coords is wrong", () => {
+        const guess: PerseusGraphType = {
+            type: "point",
+            coords: [[9, 9]],
+        };
+        const rubric: PerseusInteractiveGraphRubric = createRubric({
+            type: "point",
+            coords: [[0, 0]],
+        });
+
+        const result = InteractiveGraph.widget.validate(guess, rubric);
+
+        expect(result).toHaveBeenAnsweredIncorrectly();
+    });
+
+    it("awards points if guess.coords is right", () => {
+        const guess: PerseusGraphType = {
+            type: "point",
+            coords: [[7, 8]],
+        };
+        const rubric: PerseusInteractiveGraphRubric = createRubric({
+            type: "point",
+            coords: [[7, 8]],
+        });
+
+        const result = InteractiveGraph.widget.validate(guess, rubric);
+
+        expect(result).toEqual({
+            type: "points",
+            earned: 1,
+            total: 1,
+            message: null,
+        });
+    });
+
+    it("allows points to be specified in any order", () => {
+        const guess: PerseusGraphType = {
+            type: "point",
+            coords: [
+                [7, 8],
+                [5, 6],
+            ],
+        };
+        const rubric: PerseusInteractiveGraphRubric = createRubric({
+            type: "point",
+            coords: [
+                [5, 6],
+                [7, 8],
+            ],
+        });
+
+        const result = InteractiveGraph.widget.validate(guess, rubric);
+
+        expect(result).toHaveBeenAnsweredCorrectly();
+    });
+
+    it("does not modify the `guess` data", () => {
+        const guess: PerseusGraphType = {
+            type: "point",
+            coords: [
+                [7, 8],
+                [5, 6],
+            ],
+        };
+        const rubric: PerseusInteractiveGraphRubric = createRubric({
+            type: "point",
+            coords: [
+                [5, 6],
+                [7, 8],
+            ],
+        });
+
+        const guessClone = clone(guess);
+
+        InteractiveGraph.widget.validate(guess, rubric);
+
+        expect(guess).toEqual(guessClone);
+    });
+
+    it("does not modify the `rubric` data", () => {
+        const guess: PerseusGraphType = {
+            type: "point",
+            coords: [
+                [7, 8],
+                [5, 6],
+            ],
+        };
+        const rubric: PerseusInteractiveGraphRubric = createRubric({
+            type: "point",
+            coords: [
+                [5, 6],
+                [7, 8],
+            ],
+        });
+
+        const rubricClone = clone(rubric);
+
+        InteractiveGraph.widget.validate(guess, rubric);
+
+        expect(rubric).toEqual(rubricClone);
     });
 });